requirements.txt
streamlit>=1.32.0
requests>=2.31.0
web3>=6.4.0
websockets>=11.0.3
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
gql>=3.4.0
aiohttp>=3.9.0
colorlog>=6.7.0
plotly>=5.18.0
